refactor: moved ViewController to separate file
